+++ src/db/region.rs
@@ -7,7 +7,7 @@ use super::Pool;
 
 #[derive(Debug, Clone, Serialize, FromRow)]
 pub struct Region {
-    pub id: i64,
+    pub id: i32,
     pub device_id: String,
     pub name: String,
     pub parent_id: Option<i32>,
@@ -41,7 +41,7 @@ pub async fn list_all(pool: &Pool) -> sqlx::Result<Vec<Region>> {
     .await
 }
 
-pub async fn get_by_id(pool: &Pool, id: i64) -> sqlx::Result<Option<Region>> {
+pub async fn get_by_id(pool: &Pool, id: i32) -> sqlx::Result<Option<Region>> {
     #[cfg(feature = "mysql")]
     return sqlx::query_as::<_, Region>(
         "SELECT id, device_id, name, parent_id, parent_device_id, create_time, update_time FROM wvp_common_region WHERE id = ?",
@@ -163,7 +163,7 @@ pub async fn update(
     Ok(r.rows_affected())
 }
 
-pub async fn delete_by_id(pool: &Pool, id: i64) -> sqlx::Result<u64> {
+pub async fn delete_by_id(pool: &Pool, id: i32) -> sqlx::Result<u64> {
     #[cfg(feature = "mysql")]
     let r = sqlx::query("DELETE FROM wvp_common_region WHERE id = ?")
         .bind(id)
